[Simon LUO] Introduce more vo to support feature

+++ simoncat-taobao-front/src/main/java/com/simoncat/front/service/RecommendationDetailService.java
@@ -4,5 +4,5 @@ import com.simoncat.front.vo.RecommendationDetailVo;
 
 public interface RecommendationDetailService {
 
-	RecommendationDetailVo load(int recommendationId);
+	RecommendationDetailVo load(long recommendationId);
 }
